feat: add request for second factor at login if 2FA enabled
This commit adds verification of second factor when attempting to sign
in, if enabled.
It adds a few conditions when checking if a user is authenticated,
making that a user that has second factor enabled is authenticated
only after the second factor is authenticated.
So now, we'll still be storing the `user_id` in the session, after users
sign in, but we'll also store the boolean `user_authenticated`, which
will be true if user's second factor its authenticated.
The flow for authenticating a credential is pretty similar to the one
that we did for registering them, with the difference that we want to
fetch an existing one instead of creating it, therefore we use methods
like `WebAuthn::Credential#options_for_get` and
`WebAuthn::Credential#from_get`.