feat: discourage creation of discoverable credentials (#101)
Discoverable credentials are a special kind of webauthn credentials that
can be used for authentication without the need for the relying party to
provide any credential ids – that is, the user does not need to be
identified first in order to use the credential. This enables users to
authenticate by selecting an account from a list of discoverable
credentials they have for the current website.

However, such property of discoverable credentials is possible because
for those kind of keys, the authenticator stores the `privateKey`,
credential id and other metadata on its internal memory. Therefore, the
amount of discoverable credentials that an authenticator can create is
not unlimited (in contrast with non-discoverable credentials).

As for this demo we don't really need discoverable credentials, as we
will always have the user identified when starting the webauthn
authentication because we are using them as a second factor
authentication method, let's discourage the creation of discoverable
credentials in order to avoid consuming space in users' authenticator
unnecessarily.

References:
- https://github.com/orgs/community/discussions/49883
- https://fy.blackhats.net.au/blog/2023-02-02-how-hype-will-turn-your-security-key-into-junk/